refactor(vhdl-mac)!: move fixed point mac to nn.fixed_point
The previous location was in fact introducing
a cyclic dependency between `nn` and `vhdl`
package.

I think we will have to reconsider the architecture
but we should definitely avoid that cycle at the
current state.

Related: #461
Closes: #454

BREAKING CHANGE: will impact everything that was
importing from elasticai.creator.vhdl.shared_designs.mac
