+++ elasticai/creator/nn/fixed_point/mac/fxp_mac.tpl.vhd
