+++ elasticai/creator/nn/fixed_point/mac/mactestbench.py
@@ -1,6 +1,7 @@
 from elasticai.creator.file_generation.savable import Path
 from elasticai.creator.file_generation.template import InProjectTemplate
-from elasticai.creator.nn.fixed_point.number_converter import NumberConverter
+
+from ..number_converter import NumberConverter
 
 
 class MacTestBench:
