+++ tests/integration_tests/nn/fixed_point/mac/hw_test.py
@@ -3,9 +3,9 @@ from math import ceil, floor
 import pytest
 import torch
 
+from elasticai.creator.nn.fixed_point.mac.layer import MacLayer
 from elasticai.creator.nn.fixed_point.number_converter import FXPParams
 from elasticai.creator.vhdl.ghdl_simulation import GHDLSimulator
-from elasticai.creator.vhdl.shared_designs.mac.fixed_point.layer import MacLayer
 
 integer_test_data = [
     (FXPParams(4, 0), x1, x2)
